Fix links, dump all output files in compilation metrics page (#48)
With the changes to make tlparse artifacts unique, we also made it a bit
harder to link to compilation artifacts from pages other than
index.html. This change factors out compilationmetrics so that you can
pass the entire directory into tlparse, so that each compilation metrics
page wlil have all of the build products associated with that compile id
directly linked.

It also allows the failures and restarts page to link correctly to the
compilation metrics page, using a similar technique.

It's a bit hacky how I derive the filenames for
compilation_metrics.html, but I couldn't really come up with that good
of a way otherwise. We might wanna refactor later so that each file has
a unique id associated with it in the FXIndexMap or something.

<img width="1333" alt="image"
src="https://github.com/ezyang/tlparse/assets/4811293/35c4a830-49ab-490e-bb65-525fc1efe13b">